Burnside, numerical differentiation, reverse algebraic interpolation, modified algebraic interpolation by Newton, reverse polish notation restored, first try of embedding python with C.

+++ Project1/Variables.h
@@ -4,11 +4,11 @@
 typedef struct variable
 {
 	char* name;
-	int value;
+	double value;
 } variable;
 
 variable* variable_alloc();
-variable* variable_init(char* name, int value);
+variable* variable_init(char* name, double value);
 int variable_compare_name(const variable* v1, const variable* v2);
 int variable_compare(const variable* v1, const variable* v2);
 void variable_free(variable* v);
